+++ src/client/datascience/types.ts
@@ -843,3 +843,10 @@ type WebViewViewState = {
     readonly active: boolean;
 };
 export type WebViewViewChangeEventArgs = { current: WebViewViewState; previous: WebViewViewState };
+
+export interface INotebookProvider {
+    /**
+     * Gets or creates a notebook, and manages the lifetime of notebooks.
+     */
+    getNotebook(server: INotebookServer, resource: Uri, options?: nbformat.INotebookMetadata): Promise<INotebook>;
+}
